refactor(api): check pdf certificate download user in controller
