+++ api/src/certification/results/domain/usecases/get-certification-attestation.js
@@ -2,24 +2,12 @@
  * @typedef {import ('../../domain/usecases/index.js').CertificateRepository} CertificateRepository
  * @typedef {import ('../../domain/usecases/index.js').CertificationCourseRepository} CertificationCourseRepository
  */
-import { UnauthorizedError } from '../../../../shared/application/http-errors.js';
 
 /**
  * @param {Object} params
  * @param {CertificateRepository} params.certificateRepository
  * @param {CertificationCourseRepository} params.certificationCourseRepository
  */
-export const getCertificationAttestation = async function ({
-  userId,
-  certificationCourseId,
-  certificateRepository,
-  certificationCourseRepository,
-}) {
-  const certificationCourse = await certificationCourseRepository.get({ id: certificationCourseId });
-
-  if (certificationCourse.getUserId() !== userId) {
-    throw new UnauthorizedError();
-  }
-
+export const getCertificationAttestation = async function ({ certificationCourseId, certificateRepository }) {
   return certificateRepository.getCertificationAttestation({ certificationCourseId });
 };
